+++ data-raw/import-dataset.R
@@ -1,6 +1,7 @@
 ## code to prepare `tipnet_db` dataset goes here
 library(tidyverse)
 library(lubridate)
+library(janitor)
 library(here)  # it should follow the lubridate package because `here()`
 
 ## Raw data will stay off-line, commonly onle level above the the
@@ -50,19 +51,36 @@ import_problems <- raw_data_list %>%
 # not see any issue, anyway, it don't seams to be a real problem
 import_problems
 
+fun_double <- function(x) {all(assertive::is_whole_number(x))}
 
+# DB part of the script ------------------------------------------------
+# From this part of the script I will perform an essential cleaning
+# of the dataframes
+tipnet_db <- map(
+    raw_data_list,
+        ~ .x %>%
+            # Clean the names of the variables
+            clean_names() %>%
+            # Transform dummy variables coded as "Sì" and "No" into
+            # logical
+            mutate_if(is.character, str_to_lower) %>%
+            mutate_if(is.character, ~ . == "si")
+            # mutate_if(
+            #     fun_double, as.integer
+            # )
+    )
 
-## furter processing of the data............................................
-stop("
-    Furter preprocessing needed?
-    If so did it and then save the file as indicated"
-)
+## furter processing of the data........................................
+# stop("
+#     Furter preprocessing needed?
+#     If so did it and then save the file as indicated"
+# )
 
 
 ## Data should not stay on-line on GitHub for both privacy and
 ## space. This assure we can track the import and the general
 ## manipulation of the raw dataset while maintaining them off-line.
 
-# save(tipnet_db, file = here::here("..", "data", "tipnet_db.rda"))
+save(tipnet_db, file = here::here("..", "data", "tipnet_db.rda"))
 
 ## Please, document any specification. report any changes on `R/data.R`